Move commons and subprocess to pkg/internal
