+++ pkg/subprocess/main.go
@@ -19,7 +19,7 @@ import (
 	"github.com/hashicorp/go-hclog"
 	"github.com/hashicorp/go-plugin"
 
-	"github.com/klippa-app/go-pdfium/pkg/commons"
+	"github.com/klippa-app/go-pdfium/pkg/internal/commons"
 )
 
 func init() {
